New detect_api for profiles and global.conf (#14578)
* wip

* fixes

* more fixes

* wip

* fix

* new test

* global template too
