+++ conans/test/unittests/client/conf/detect/test_gcc_compiler.py
@@ -3,8 +3,8 @@ import unittest
 import mock
 from parameterized import parameterized
 
-from conans.client.conf.detect import _gcc_compiler
-from conan.api.output import ConanOutput
+
+from conan.internal.api.detect_api import _gcc_compiler
 
 
 class GCCCompilerTestCase(unittest.TestCase):
@@ -12,7 +12,7 @@ class GCCCompilerTestCase(unittest.TestCase):
     @parameterized.expand([("10",), ("4.2",), ('7', )])
     def test_detect_gcc_10(self, version):
         with mock.patch("platform.system", return_value="Linux"):
-            with mock.patch("conans.client.conf.detect.detect_runner", return_value=(0, version)):
-                compiler, installed_version = _gcc_compiler(ConanOutput())
+            with mock.patch("conan.internal.api.detect_api.detect_runner", return_value=(0, version)):
+                compiler, installed_version = _gcc_compiler()
         self.assertEqual(compiler, 'gcc')
         self.assertEqual(installed_version, version)
