+++ conans/test/unittests/util/detected_architecture_test.py
@@ -1,14 +1,9 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# vim: tabstop=8 expandtab shiftwidth=4 softtabstop=4
-
-
 import mock
 import unittest
 
 from parameterized import parameterized
 
-from conans.client.conf.detect import _detected_architecture
+from conan.internal.api.detect_api import detect_arch
 
 
 class DetectedArchitectureTest(unittest.TestCase):
@@ -38,22 +33,22 @@ class DetectedArchitectureTest(unittest.TestCase):
     def test_various(self, mocked_machine, expected_arch):
 
         with mock.patch("platform.machine", mock.MagicMock(return_value=mocked_machine)):
-            self.assertEqual(expected_arch, _detected_architecture(), "given '%s' expected '%s'" % (mocked_machine, expected_arch))
+            self.assertEqual(expected_arch, detect_arch(), "given '%s' expected '%s'" % (mocked_machine, expected_arch))
 
     def test_aix(self):
         with mock.patch("platform.machine", mock.MagicMock(return_value='00FB91F44C00')),\
                 mock.patch("platform.processor", mock.MagicMock(return_value='powerpc')),\
                 mock.patch("platform.system", mock.MagicMock(return_value='AIX')),\
-                mock.patch("conans.client.conf.detect._get_aix_conf", mock.MagicMock(return_value='32')),\
+                mock.patch("conan.internal.api.detect_api._get_aix_conf", mock.MagicMock(return_value='32')),\
                 mock.patch('subprocess.check_output', mock.MagicMock(return_value='7.1.0.0')):
-            self.assertEqual('ppc32', _detected_architecture())
+            self.assertEqual('ppc32', detect_arch())
 
         with mock.patch("platform.machine", mock.MagicMock(return_value='00FB91F44C00')),\
                 mock.patch("platform.processor", mock.MagicMock(return_value='powerpc')),\
                 mock.patch("platform.system", mock.MagicMock(return_value='AIX')),\
-                mock.patch("conans.client.conf.detect._get_aix_conf", mock.MagicMock(return_value='64')),\
+                mock.patch("conan.internal.api.detect_api._get_aix_conf", mock.MagicMock(return_value='64')),\
                 mock.patch('subprocess.check_output', mock.MagicMock(return_value='7.1.0.0')):
-            self.assertEqual('ppc64', _detected_architecture())
+            self.assertEqual('ppc64', detect_arch())
 
     def test_solaris(self):
         with mock.patch("platform.machine", mock.MagicMock(return_value='sun4v')),\
@@ -61,14 +56,14 @@ class DetectedArchitectureTest(unittest.TestCase):
                 mock.patch("platform.system", mock.MagicMock(return_value='SunOS')),\
                 mock.patch("platform.architecture", mock.MagicMock(return_value=('64bit', 'ELF'))),\
                 mock.patch("platform.release", mock.MagicMock(return_value='5.11')):
-            self.assertEqual('sparcv9', _detected_architecture())
+            self.assertEqual('sparcv9', detect_arch())
 
         with mock.patch("platform.machine", mock.MagicMock(return_value='i86pc')),\
                 mock.patch("platform.processor", mock.MagicMock(return_value='i386')),\
                 mock.patch("platform.system", mock.MagicMock(return_value='SunOS')),\
                 mock.patch("platform.architecture", mock.MagicMock(return_value=('64bit', 'ELF'))),\
                 mock.patch("platform.release", mock.MagicMock(return_value='5.11')):
-            self.assertEqual('x86_64', _detected_architecture())
+            self.assertEqual('x86_64', detect_arch())
 
     @parameterized.expand([
         ["E1C+", "e2k-v4"],
@@ -85,4 +80,4 @@ class DetectedArchitectureTest(unittest.TestCase):
     def test_e2k(self, processor, expected_arch):
         with mock.patch("platform.machine", mock.MagicMock(return_value='e2k')), \
                 mock.patch("platform.processor", mock.MagicMock(return_value=processor)):
-            self.assertEqual(expected_arch, _detected_architecture())
+            self.assertEqual(expected_arch, detect_arch())
